pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "power-data-platform-poc"
version = "0.1.0"
description = "PoC data platform for Japanese public power system data."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "polars>=1.18.0,<2.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "pydantic>=2.10.0,<3.0.0",
  "pyyaml>=6.0.2,<7.0.0",
  "python-dateutil>=2.9.0,<3.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "playwright>=1.52.0,<2.0.0",
  "uvicorn>=0.34.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "ruff>=0.9.0,<1.0.0",
]

[project.scripts]
power-poc-ingest = "pipelines.ingest:main"
power-poc-api = "apps.api.main:run"

[tool.setuptools]
py-modules = ["config"]

[tool.setuptools.packages.find]
include = [
  "apps*",
  "collectors*",
  "loaders*",
  "parsers*",
  "pipelines*",
  "validators*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
